requirements.txt
# Azure AI SDK - Core dependencies
azure-ai-projects>=1.0.0
azure-identity>=1.15.0

# Environment management
python-dotenv>=1.0.0

# Optional: For extending with weather or other API functionality
requests>=2.31.0
